fixed uname change + acc del

+++ php_includes/settings_common.php
@@ -208,7 +208,7 @@
   }
 
   function remDir($path) {
-    if(is_dir($path)){
+    if(is_dir($path) && file_exists($path)){
       rmdir($path);
     }
   }
